feat: Implement user detail and list components with loading skeletons
- Added UserDetailComponent for displaying user details and tasks.
- Created UserDetailPageWrapper to encapsulate UserDetailComponent.
- Introduced UserDetailSkeleton for loading state representation.
- Updated loading.tsx to use UserDetailSkeleton.
- Refactored page.tsx to utilize UserDetailPageWrapper with suspense.
- Developed UserListComponent for displaying a list of users with pagination and search.
- Added UserListSkeleton for loading state in user list.
- Refactored users page.tsx to use UserListComponent with suspense.
