+++ src/app/(dashboard)/admin/users/[id]/components/UserDetailPageWrapper.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import UserDetailPageClient from './UserDetailComponent';
+
+export default function UserDetailPageWrapper({ userId }: { userId: string }) {
+  return <UserDetailPageClient userId={userId} />;
+}
