+++ src/app/(dashboard)/admin/users/[id]/loading.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import UserDetailSkeleton from './components/UserDetailSkeleton';
+
+export default function Loading() {
+  return <UserDetailSkeleton />;
+}
